### 2025-07-21 #8625 @editorialbot
[PRE REVIEW]: ConversationAlign: An R package for Computing Linguistic Alignment and Corpus Analytics in Dyadic Conversation Transcripts
**Submitting author:** @reilly-lab (<a href="http://orcid.org/0000-0002-0891-438X">Jamie Reilly</a>)
**Repository:** https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign
**Branch with paper.md** (empty if default branch): main
**Version:** v0.2.0
