### 2025-07-21 #8625 @reilly-lab
Thank you for considering our paper. Here are some potential reviewers:

Potential reviewers:
Sebastian Speer  ***@***.******@***.***>
Emma Templeton  ***@***.***
Elen Le Foll  ***@***.******@***.***>
Ray Becker  ***@***.******@***.***>

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Jamie Reilly, Ph.D.
Professor, Department of Communication Sciences & Disorders
Eleanor M. Saffran Center for Cognitive Neuroscience
Ritter Annex 995 (office);  Ritter Annex 999 (lab)
1701 N. 13th Street, Weiss Hall 110
Philadelphia, PA 19122
Tel:     215.204.3995
Lab:  www.reilly-coglab.com<http://www.reilly-coglab.com/>

This communication may contain information that is legally protected from unauthorized disclosure.  If you are not the intended recipient, please note that any dissemination, distribution or copying of this communication is strictly prohibited.  If you have received this message in error, you should notify the sender immediately by telephone or by return email and delete this message from your computer.

From: The Open Journals editorial robot ***@***.***>
Date: Monday, July 21, 2025 at 6:05 AM
Cc: Jamie Reilly ***@***.***>, Mention ***@***.***>
Subject: [External] [openjournals/joss-reviews] [PRE REVIEW]: ConversationAlign: An R package for Computing Linguistic Alignment and Corpus Analytics in Dyadic Conversation Transcripts (Issue #8625)
[Image removed by sender.]editorialbot created an issue (openjournals/joss-reviews#8625)<https://github.com/openjournals/joss-reviews/issues/8625>

Submitting author: @reilly-lab<https://github.com/reilly-lab> (Jamie Reilly<http://orcid.org/0000-0002-0891-438X>)
Repository: https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign
Branch with paper.md (empty if default branch): main
Version: v0.2.0
Editor: Pending
Reviewers: Pending
Managing EiC: Samuel Forbes

Status

[Image removed by sender. status]<https://joss.theoj.org/papers/839ab720504a5c966b4b2893f78ec2b2>

Status badge code:

Author instructions

Thanks for submitting your paper to JOSS @reilly-lab<https://github.com/reilly-lab>. Currently, there isn't a JOSS editor assigned to your paper.

@reilly-lab<https://github.com/reilly-lab> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search the list of people<https://reviewers.joss.theoj.org/lookup> that have already agreed to review and may be suitable for this submission.

Editor instructions

The JOSS submission bot @editorialbot<https://github.com/editorialbot> is here to help you find and assign reviewers and start the main review. To find out what @editorialbot<https://github.com/editorialbot> can do for you type:
