### 2025-07-21 5b09355
Added JOSS submission status badge to readme
