### 2025-07-21 129f381
Increment version number to 0.3.0
