### 2025-07-21 b737370
changed software license and removed JOSS paper from gitignore and rbuildignore
