+++ paper/paper.bib
@@ -1,0 +1,77 @@
+@book{pickering_understanding_2021,
+    title = {Understanding Dialogue: Language Use and Social Interaction},
+    shorttitle = {Understanding Dialogue},
+    url = {https://books.google.com/books?hl=en&lr=&id=3RgXEAAAQBAJ&oi=fnd&pg=PR7&dq=pickering+garrod+understanding+dialogue&ots=0qe68OV8Xs&sig=ulM_ibE3lLJewbmVg-UvQvfEHhM},
+    urldate = {2025-06-26},
+    publisher = {Cambridge University Press},
+    author = {Pickering, Martin J. and Garrod, Simon},
+    year = {2021}
+}
+
+@article{benoit_quanteda_2018,
+    title = {quanteda: An R Package for the Quantitative Analysis of Textual Data},
+    volume = {3},
+    copyright = {http://creativecommons.org/licenses/by/4.0/},
+    issn = {2475-9066},
+    shorttitle = {quanteda},
+    url = {https://doi.org/10.21105/joss.00774},
+    doi = {10.21105/joss.00774},
+    number = {30},
+    urldate = {2025-01-05},
+    journal = {Journal of Open Source Software},
+    author = {Benoit, Kenneth and Watanabe, Kohei and Wang, Haiyan and Nulty, Paul and Obeng, Adam and Müller, Stefan and Matsuo, Akitaka},
+    month = oct,
+    year = {2018},
+    pages = {774}
+}
+
+@misc{michalke_korpus_2018,
+    title = {{koRpus}: An R Package for Text Analysis},
+    copyright = {GPL (≥ 3)},
+    shorttitle = {{koRpus}},
+    url = {https://CRAN.R-project.org/package=koRpus},
+    abstract = {A set of tools to analyze texts. Includes, amongst others, functions for automatic language detection, hyphenation, several indices of lexical diversity (e.g., type token ratio, HD-D/vocd-D, MTLD) and readability (e.g., Flesch, SMOG, LIX, Dale-Chall). Basic import functions for language corpora are also provided, to enable frequency analyses (supports Celex and Leipzig Corpora Collection file formats) and measures like tf-idf. Note: For full functionality a local installation of TreeTagger is recommended. It is also recommended to not load this package directly, but by loading one of the available language support packages from the 'l10n' repository {\textless}https://undocumeantit.github.io/repos/l10n{\textgreater}. 'koRpus' also includes a plugin for the R GUI and IDE RKWard, providing graphical dialogs for its basic features. The respective R package 'rkward' cannot be installed directly from a repository, as it is a part of RKWard. To make full use of this feature, please install RKWard from {\textless}https://rkward.kde.org{\textgreater} (plugins are detected automatically). Due to some restrictions on CRAN, the full package sources are only available from the project homepage. To ask for help, report bugs, request features, or discuss the development of the package, please subscribe to the koRpus-dev mailing list ({\textless}http://korpusml.reaktanz.de{\textgreater}).},
+    urldate = {2019-12-23},
+    author = {Michalke, Meik and Brown, Earl and Mirisola, Alberto and Brulet, Alexandre and Hauser, Laura},
+    month = oct,
+    year = {2018},
+    keywords = {NaturalLanguageProcessing}
+}
+
+@article{duran_align_2019,
+    title = {{ALIGN}: Analyzing Linguistic Interactions with Generalizable techNiques—A Python Library},
+    volume = {24},
+    issn = {1939-1463},
+    shorttitle = {{ALIGN}},
+    url = {https://doi.org/10.1037/met0000206},
+    doi = {10.1037/met0000206},
[…]
