+++ paper/paper.md
@@ -1,0 +1,104 @@
+---
+title: 'ConversationAlign: An R package for Computing Linguistic Alignment and Corpus Analytics in Dyadic Conversation Transcripts'
+tags:
+  - R
+  - conversation analysis
+  - alignment
+  - psycholinguistics
+  - natural language processing
+authors:
+  - name: Jamie Reilly
+    orcid: 0000-0002-0891-438X
+    equal-contrib: false
+    corresponding: true
+    affiliation: "1, 2"
+    
+  - name: Benjamin Sacks
+    orcid: 0009-0005-7323-6625 
+    equal-contrib: true
+    corresponding: false
+    affiliation: 2
+    
+  - name: Virginia Ulichney
+    orcid: 0000-0002-8170-1313
+    equal-contrib: false
+    corresponding: false
+    affiliation: 2
+    
+  - name: Gus Cooney
+    orcid: 0000-0003-4082-7515 
+    equal-contrib: false
+    corresponding: false
+    affiliation: "3"
+    
+  - name: Chelsea Helion
+    orcid: 0000-0002-3149-1493
+    equal-contrib: false
+    corresponding: false
+    affiliation: "1,2"
+    
+affiliations:
+  - name: Department of Communication Sciences and Disorders, Temple University, United States
+    index: 1
+  - name: Department of Psychology and Neuroscience, Temple University, United States
+    index: 2
+  - name: Wharton School, University of Pennsylvania, United States
+    index: 3
+
+correspondence: jamie.reilly@temple.edu
+date: " 9 July 2025"
+bibliography: paper.bib
+abstract: |
+ `ConversationAlign` is an R package that executes a series of operations upon one or more conversation transcripts (i.e., two-person dialogues). Transcripts nominally contain at least two variables (speaker  identity and text).`ConversationAlign` will retain all other meta-data such as timestamps, demographics, and grouping variables.  `ConversationAlign` imports raw transcripts into R, appends unique document identifiers, and concatenates all conversations into a single dataframe. `ConversationAlign` generates corpus analytics characterizing the conversation transcript(s) of interest. Users guide a number of text cleaning operations such as stopword removal and lemmatization. The package ultimately vectorizes the original text into a one-word-per-row format. `ConversationAlign` yokes published norms to each content word spanning more than 40 lexical, affective, and semantic dimensions (e.g. word length, morphological complexity, arousal, valence). `ConversationAlign` outputs summary data for each conversation including main effects and indices of local and global alignment for each specified dimension of interest.   
+keywords: |
+  conversation analysis; discourse; language processing; alignment
+authorcontributions: |
+  JR, VU, and BS conceived the software. All authors drafted and edited the paper.
+funding: |
+  This research was funded by R01 DC013063 from the US National Institute on Deafness and Other Communicative Disorders (NIH/NIDCD).
+conflictsofinterest: |
+  The authors declare no conflicts of interest.
+output: rticles::joss_article 
+keep_tex: true
+csl: apa.csl
+journal: JOSS
+---
+
+# Statment of Need
[…]
