+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   CA_Step2_Prep: CA_Step2_Prep.html
   CA_Step3_Summarize: CA_Step3_Summarize.html
   CA_Step4_Analytics: CA_Step4_Analytics.html
-last_built: 2025-07-21T15:47Z
+last_built: 2025-07-21T16:19Z
 urls:
   reference: https://reilly-conceptscognitionlab.github.io/ConversationAlign/reference
   article: https://reilly-conceptscognitionlab.github.io/ConversationAlign/articles
