### 2025-07-21 42d8acc
Rebuilt site, deleted old vignettes
