### 2025-07-21 #8625 @reilly-lab
> **Paper file info**:
>
> ⚠️ Failed to find a paper file in https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign (branch: main)

Removed all references to paper from gitignore and RBuildignore, moved paper from inst folder to main branch. Paper and corresponding bibtext files should appear in root. Incremented package version to 0.3.0
