### 2025-07-21 #8625 @reilly-lab
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/16414070936?check_suite_focus=true). Paper file not found.

We have moved the file/folder from `inst` to the the main branch. All should be visible now.
