### 2025-07-21 #8625 @reilly-lab
> Checking the BibTeX entries failed with the following error:
>
> ```
> No paper file path
> ```

Moved paper to main branch from `inst`
