+++ pkgdown.yml
@@ -2,12 +2,12 @@ pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
-  CA_Intro: CA_Intro.html
-  CA_Step1_Read: CA_Step1_Read.html
-  CA_Step2_Prep: CA_Step2_Prep.html
-  CA_Step3_Summarize: CA_Step3_Summarize.html
-  CA_Step4_Analytics: CA_Step4_Analytics.html
-last_built: 2025-07-21T16:59Z
+  ConversationAlign Introduction: ConversationAlign Introduction.html
+  ConversationAlign Step1 Read: ConversationAlign Step1 Read.html
+  ConversationAlign Step2 Prep: ConversationAlign Step2 Prep.html
+  ConversationAlign Step3 Summarize: ConversationAlign Step3 Summarize.html
+  ConversationAlign Step4 Analytics: ConversationAlign Step4 Analytics.html
+last_built: 2025-07-21T17:11Z
 urls:
   reference: https://reilly-conceptscognitionlab.github.io/ConversationAlign/reference
   article: https://reilly-conceptscognitionlab.github.io/ConversationAlign/articles
