### 2025-07-22 #8625 @reilly-lab
> Checking the BibTeX entries failed with the following error:
>
> ```
> No paper file path
> ```

Moved paper and bib to `main` from `inst` -- everything should be accessible now. Apologies.
