### 2025-07-27 85f2f1c
updated statement of need in JOSS paper
