+++ paper/paper.md
@@ -46,10 +46,10 @@ affiliations:
     index: 3
 
 correspondence: jamie.reilly@temple.edu
-date: " 9 July 2025"
+date: "27 July 2025"
 bibliography: paper.bib
 abstract: |
- `ConversationAlign` is an R package that executes a series of operations upon one or more conversation transcripts (i.e., two-person dialogues). Transcripts nominally contain at least two variables (speaker  identity and text).`ConversationAlign` will retain all other meta-data such as timestamps, demographics, and grouping variables.  `ConversationAlign` imports raw transcripts into R, appends unique document identifiers, and concatenates all conversations into a single dataframe. `ConversationAlign` generates corpus analytics characterizing the conversation transcript(s) of interest. Users guide a number of text cleaning operations such as stopword removal and lemmatization. The package ultimately vectorizes the original text into a one-word-per-row format. `ConversationAlign` yokes published norms to each content word spanning more than 40 lexical, affective, and semantic dimensions (e.g. word length, morphological complexity, arousal, valence). `ConversationAlign` outputs summary data for each conversation including main effects and indices of local and global alignment for each specified dimension of interest.   
+ `ConversationAlign` is an R package that executes a series of operations upon one or more conversation transcripts (i.e., two-person dialogues). Transcripts nominally contain at least two variables (speaker identity and text). In addition to these essential fields, `ConversationAlign` will retain all other meta-data such as timestamps, demographics, and grouping variables.  `ConversationAlign` imports raw transcripts into R, appends unique document identifiers, and concatenates all conversations into a single dataframe. `ConversationAlign` generates corpus analytics characterizing the conversation transcript(s) of interest. Users guide a number of text cleaning operations such as stopword removal and lemmatization. The package ultimately vectorizes the original text into a one-word-per-row format. `ConversationAlign` yokes published norms to each content word spanning more than 40 lexical, affective, and semantic dimensions (e.g. word length, morphological complexity, arousal, valence). `ConversationAlign` outputs summary data for each conversation including main effects and indices of local and global alignment for each specified dimension of interest.   
 keywords: |
   conversation analysis; discourse; language processing; alignment
 authorcontributions: |
@@ -65,16 +65,18 @@ journal: JOSS
 ---
 
 # Statment of Need
-Conversation is among the most complex social and linguistic behaviors that humans routinely undertake. In a dyadic interaction, two conversation partners modify the form and content of their own production to align with each other [@pickering_understanding_2021]. This process, known as linguistic alignment, occurs across many dimensions (e.g., affective coloring, prosody, gesture, formality, semantic and syntactic complexity).   `ConversationAlign` offers an automated approach to computing empirical indices alignment between conversation partners across one or more conversations with coverage of over 40 distinct psycholinguistic dimensions (e.g., word length, valence, concreteness, morphological complexity).  Although other open-source software applications exist for text corpus analyses, such as Quanteda [@benoit_quanteda_2018] and Korpus [@michalke_korpus_2018], we know of R-specific packages that offer a comprehensive text processing pipeline capable of cleaning, formatting, transforming, and summarizing raw conversation transcripts (but for Python see ALIGN [@duran_align_2019]).
+Although many excellent text analysis applications exist (e.g., Quanteda [@benoit_quanteda_2018] and Korpus [@michalke_korpus_2018]), we know of no R packages that are tailored to the unique demands of conversation analysis (but for Python see ALIGN [@duran_align_2019]). `ConversationAlign` offers a comprehensive text processing pipeline and novel algoirthms for computing linguistic alignment in 2-person dialogues. This software offers standardization and automation advantages that are in great need in a field that has historically relied heavily upon manual coding systems and subjective human judgment.
 
[…]
