### 2025-07-27 #8625 @reilly-lab
> **Paper file info**:
>
> ⚠️ Wordcount for `paper.md` is **1211**
>
> 🔴 Failed to discover a `Statement of need` section in paper

Added a `statement of need` section and trimmed the length of the paper
