### 2025-07-29 14c2f94
Increment version number to 0.3.2
