### 2025-07-29 c9e5471
updated news and incremented package to 0.3.2
