+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   ConversationAlign Step2 Prep: ConversationAlign Step2 Prep.html
   ConversationAlign Step3 Summarize: ConversationAlign Step3 Summarize.html
   ConversationAlign Step4 Analytics: ConversationAlign Step4 Analytics.html
-last_built: 2025-07-27T20:55Z
+last_built: 2025-07-29T19:00Z
 urls:
   reference: https://reilly-conceptscognitionlab.github.io/ConversationAlign/reference
   article: https://reilly-conceptscognitionlab.github.io/ConversationAlign/articles
