### 2025-09-18 1602a01
add lag of one to dAUC function
