+++ R/summarize_dyads.R
@@ -116,12 +116,12 @@ summarize_dyads <- function(df_prep, custom_lags = NULL, sumdat_only = TRUE, cor
   auc_df_long <- auc_df %>%
     tidyr::pivot_longer(
       contains("AUC"),
-      names_to = c("Dimension", "reshaped"),
-      names_pattern = "AUC_(.*)_(raw|scaled100)",
+      names_to = c("Dimension", "reshaped", "Lag"),
+      names_pattern = "AUC_(.*)_(raw|scaled100)_(Immediate|Lag\\d|Lead\\d)",
       values_to = "AUC"
     ) %>%
-    tidyr::pivot_wider(
-      names_from = reshaped,
+    tidyr::pivot_wider( # pivot out the reshaped and lag columns
+      names_from = c(reshaped, Lag),
       names_prefix = "AUC_",
       values_from = AUC
     )
