### 2025-09-20 b936185
update tests for summarize dyads and remove unassigned global variable note in summarize_dyads
