+++ DESCRIPTION
@@ -41,7 +41,7 @@ Suggests:
     rmarkdown,
     testthat (>= 3.0.0)
 URL: https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 LazyData: true
 VignetteBuilder: knitr
 Collate: 
