+++ NAMESPACE
@@ -33,7 +33,6 @@ importFrom(purrr,map_dfr)
 importFrom(rlang,":=")
 importFrom(rlang,sym)
 importFrom(stats,cor.test)
-importFrom(stats,lag)
 importFrom(stats,na.omit)
 importFrom(stats,sd)
 importFrom(stringi,stri_replace_all_fixed)
