### 2025-09-26 2c3e914
add generate_shams()
