+++ man/generate_shams.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/generate_shams.R
+\name{generate_shams}
+\alias{generate_shams}
+\title{generate_shams}
+\usage{
+generate_shams(df_prep, seed = NULL)
+}
+\arguments{
+\item{df_prep}{Output dataframe of prep_dyads().}
+
+\item{seed}{(Optional) a seed for reproducibility in random sampling}
+}
+\value{
+A dataframe similar to prepped dyads, with each participant's time series randomly shuffled.
+}
+\description{
+Generates a permutation of each individual dyad. Shuffled dyads may act as controls to their originals.
+}
