+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(corpus_analytics)
+export(generate_shams)
 export(prep_dyads)
 export(read_1file)
 export(read_dyads)
@@ -18,6 +19,7 @@ importFrom(dplyr,lag)
 importFrom(dplyr,left_join)
 importFrom(dplyr,matches)
 importFrom(dplyr,mutate)
+importFrom(dplyr,n)
 importFrom(dplyr,n_distinct)
 importFrom(dplyr,na_if)
 importFrom(dplyr,rename)
