+++ NAMESPACE
@@ -10,6 +10,7 @@ importFrom(DescTools,AUC)
 importFrom(dplyr,across)
 importFrom(dplyr,bind_cols)
 importFrom(dplyr,bind_rows)
+importFrom(dplyr,consecutive_id)
 importFrom(dplyr,distinct)
 importFrom(dplyr,everything)
 importFrom(dplyr,filter)
