+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   ConversationAlign_Step2_Prep: ConversationAlign_Step2_Prep.html
   ConversationAlign_Step3_Summarize: ConversationAlign_Step3_Summarize.html
   ConversationAlign_Step4_Analytics: ConversationAlign_Step4_Analytics.html
-last_built: 2025-10-04T14:28Z
+last_built: 2025-10-09T22:59Z
 urls:
   reference: https://reilly-conceptscognitionlab.github.io/ConversationAlign/reference
   article: https://reilly-conceptscognitionlab.github.io/ConversationAlign/articles
