### 2025-10-15 209989a
update summarize dyads article with generate shams
