### 2025-10-15 3404c4c
update readme with generate_shams and keep_backchannel
