### 2025-10-15 959b064
update compute_auc and tests to fit new dAUC column names
