+++ tests/testthat/test-compute_auc.R
@@ -9,8 +9,8 @@ test_that("auc formats properly", {
   # computed previously
   expected_auc_df <- data.frame(
     Event_ID = "LittleLamb", Exchanges = 50, Talked_First = as.factor("Mary"),
-    AUC_emo_anger_raw_Immediate = 1.485925, AUC_emo_anger_scaled100_Immediate = 1.485925,
-    AUC_emo_anger_raw_Lag1 = 1.4556, AUC_emo_anger_scaled100_Lag1 = 1.4556
+    AUC_emo_anger_raw_Immediate = 1.485925, AUC_emo_anger_scaled50_Immediate = 1.485925,
+    AUC_emo_anger_raw_Lag1 = 1.4556, AUC_emo_anger_scaled50_Lag1 = 1.4556
   )
   expected_auc_df$Talked_First <- factor(expected_auc_df$Talked_First, levels = c("Dave", "Mary"))
   # compare to computed df
@@ -23,8 +23,8 @@ one_rhyme_small <- one_rhyme[one_rhyme$Exchange_Count < 3, ]
 test_that("auc fills with na when under 3 exchanges", {
   expected_small_auc_df <- data.frame(
     Event_ID = "LittleLamb", Exchanges = 2, Talked_First = as.factor("Mary"),
-    AUC_emo_anger_raw_Immediate = as.double(NA), AUC_emo_anger_scaled100_Immediate = as.double(NA),
-    AUC_emo_anger_raw_Lag1 = as.double(NA), AUC_emo_anger_scaled100_Lag1 = as.double(NA)
+    AUC_emo_anger_raw_Immediate = as.double(NA), AUC_emo_anger_scaled50_Immediate = as.double(NA),
+    AUC_emo_anger_raw_Lag1 = as.double(NA), AUC_emo_anger_scaled50_Lag1 = as.double(NA)
   )
   expected_small_auc_df$Talked_First <- factor(expected_small_auc_df$Talked_First, levels = c("Dave", "Mary"))
   # compare to computed df
