+++ README.Rmd
@@ -17,7 +17,7 @@ knitr::opts_chunk$set(
 <br>
 
 <h1 style="color: #8B0000;">ConversationAlign</h1>
-Open-source software for computing main effects and indices of alignment across coversation partners in dyadic conversation transcripts.
+Open-source software for computing main effects and indices of alignment across conversation partners in dyadic conversation transcripts.
 
 # <a href="https://reilly-conceptscognitionlab.github.io/ConversationAlign/"><img src="man/figures/logo.png" align="left" height="138" alt="ConversationAlign website" style="margin-right: 20px; margin-bottom: 20px;" /></a>
 
