+++ llms.txt
@@ -1,0 +1,322 @@
+# ConversationAlign
+
+  
+  
+  
+
+Open-source software for computing main effects and indices of alignment
+across coversation partners in dyadic conversation transcripts.
+
+# 
+
+[![GitHub
+release](https://img.shields.io/github/v/release/Reilly-ConceptsCognitionLab/ConversationAlign?color=blue&include_prereleases&label=Release)](https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign/releases)[![status](https://joss.theoj.org/papers/839ab720504a5c966b4b2893f78ec2b2/status.svg)](https://joss.theoj.org/papers/839ab720504a5c966b4b2893f78ec2b2)
+
+[![GitHub
+stars](https://img.shields.io/github/stars/Reilly-ConceptsCognitionLab/ConversationAlign?style=social)](https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign/stargazers)
+
+[![Maintenance](https://img.shields.io/badge/Maintained%3F-yes-green.svg)](https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign/graphs/commit-activity)
+[![License: GPL
+v3+](https://img.shields.io/badge/License-GPL%20v3+-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
+[![R-CMD-check](https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign/actions/workflows/R-CMD-check.yaml)
+
+![illustration of processing pipeline used by
+ConversationAlign](reference/figures/convo_demo.jpg)
+
+# Overview
+
+`ConversationAlign` analyzes alignment and computes main effects across
+more than 40 unique dimensions between interlocutors (conversation
+partners) engaged in two-person conversations. ConversationAlign
+transforms raw language data into simultaneous time series objects
+across \>40 possible dimensions via an embedded lookup database. There
+are a number of issues you should consider and steps you should take to
+prepare your data.  
+
+# License
+
+`ConversationAlign` is licensed under the [GNU LGPL
+v3.0](https://www.gnu.org/licenses/lgpl-3.0).  
+
+# Installation and Technical Considerations
+
+One of the main features of the `ConversationAlign` algorithm involves
+yoking norms for many different lexical, affective, and semantic
+dimensions to each content word in your conversation transcripts of
+interest. We accomplish this by joining your data to several large
+lookup databases. These databases are too large to embed within
+`ConversationAlign`. When you load `ConversationAlign`, all of these
+databases should automatically download and load from an external
+companionn repository `ConversationAlign_Data`. `ConversationAlign`
+needs these data, so you will need a decent internet connection to load
+the package. It might take a second or two to complete the download if
+Github is acting up. Install the development version of
+ConversationAlign from [GitHub](https://github.com/) using the
+`devtools` package.
+
+``` r
+# Check if devtools is installed, if not install it
+if (!require("devtools", quietly = TRUE)) {
+  install.packages("devtools")
+}
+
+# Load devtools
+library(devtools)
+
+# Check if ConversationAlign is installed, if not install from GitHub
+if (!require("ConversationAlign", quietly = TRUE)) {
+  devtools::install_github("Reilly-ConceptsCognitionLab/ConversationAlign")
+}
+
+# Load SemanticDistance
+library(ConversationAlign)
+```
+
+# Step 1: Read and Format Transcript Options
+
+## `read_dyads()`
+
+- Reads transcripts from a local drive or directory of your choice.
[…]
