### 2025-11-08 db8e3e0
repaired 404 links in vignettes
