+++ doc/CA_Intro.Rmd
@@ -81,7 +81,7 @@ str(MaronGross_2013)
 # Caveat emptor
 Any analysis of language comes with assumptions and potential bias. For example, there are some instances where a researcher might care about morphemes and grammatical elements such as 'the', 'a', 'and', etc..  The default for ConversationAlign is to omit these as stopwords and to average across all open class words (e.g., nouns, verbs) in each turn by interlocutor. There are some specific cases where this can all go wrong. Here are some things to consider: <br>
 
-1. <span style="color:red;">Stopwords </span>: ``ConversationAlign`` omits stopwords by default applying a customized stopword list, ``Temple_Stopwords25``. [CLICK HERE](https://osf.io/dc5k7) to inspect the list. This stopword list includes greetings, idioms, filler words, numerals, and pronouns.
+1. <span style="color:red;">Stopwords </span>: ``ConversationAlign`` omits stopwords by default applying a customized stopword list, ``Temple_Stopwords25``. This stopword list includes greetings, idioms, filler words, numerals, and pronouns.
 
 2. <span style="color:red;">Lemmatization </span>: The package will lemmatize your language transcripts by default. Lemmatization transforms inflected forms (e.g., standing, stands) into their root or dictionary entry (e.g., stand). This helps for yoking offline values (e.g., happiness, concreteness) to each word and also entails what NLP folks refer to as 'term aggregation'. However, sometimes you might NOT want to lemmatize. You can easily change this option by using the argument, "lemmatize=FALSE," to the clean_dyads function below. <br>
 
@@ -96,13 +96,13 @@ Any analysis of language comes with assumptions and potential bias. For example,
 
 1. **Preprint** <br>
 Our PsyArXiv preprint describing the method(s) in greater detail is referenced as:
-Sacks, B., Ulichney, V., Duncan, A., Helion, C., Weinstein, S., Giovannetti, T., ... Reilly, J. (2025, March 12). *ConversationAlign: Open-Source Software for Analyzing Patterns of Lexical Use and Alignment in Conversation Transcripts*. [Click Here](https://osf.io/preprints/psyarxiv/7xqnp_v1) to read our preprint. It was recently invited for revision at Behavior Rsearch Methods. We will update when/if eventually accepted there! <br>
+Sacks, B., Ulichney, V., Duncan, A., Helion, C., Weinstein, S., Giovannetti, T., ... Reilly, J. (2025, March 12). *ConversationAlign: Open-Source Software for Analyzing Patterns of Lexical Use and Alignment in Conversation Transcripts*. <br>
 
 2. **Methods for creating internal lookup database** <br>
-ConversationAlign contains a large, internal lexical lookup_database. [Click Here](https://reilly-lab.github.io/ConversationAlign_LookupHowTo_Jun25.html) to see how we created this by merging other offline psycholinguistic databases into one. <br>
+ConversationAlign contains a large, internal lexical lookup_database. [CLICK HERE](https://reilly-lab.github.io/ConversationAlign_LookupHowTo_Jun25.html) to see how we created this by merging other offline psycholinguistic databases into one. <br>
 
 3. **Variable Key for ConversationAlign** <br>
-ConversationAlign currently allows users to compute alignment dynamics across 30 different lexical, affective, and semantic dimensions.[Click Here](https://reilly-lab.github.io/ConversationAlign_LookupJul25_VariableKey.html) to link to a variable key. <br>
+ConversationAlign currently allows users to compute alignment dynamics across 30 different lexical, affective, and semantic dimensions.[CLICK HERE](https://reilly-lab.github.io/ConversationAlign_LookupJul25_VariableKey.html) to link to a variable key. <br>
 
 # References
 Lewis, David D., et al. (2004) "Rcv1: A new benchmark collection for text categorization research." Journal of machine learning research 5: 361-397.
