+++ articles/ConversationAlign_Introduction.md
@@ -199,23 +199,22 @@ Here are some things to consider:
     referenced as: Sacks, B., Ulichney, V., Duncan, A., Helion, C.,
     Weinstein, S., Giovannetti, T., … Reilly, J. (2025, March 12).
     *ConversationAlign: Open-Source Software for Analyzing Patterns of
-    Lexical Use and Alignment in Conversation Transcripts*. [Click
-    Here](https://osf.io/preprints/psyarxiv/7xqnp_v1) to read our
+    Lexical Use and Alignment in Conversation Transcripts*. to read our
     preprint. It was recently invited for revision at Behavior Rsearch
     Methods. We will update when/if eventually accepted there!  
 
 2.  **Methods for creating internal lookup database**  
     ConversationAlign contains a large, internal lexical
     lookup_database. [Click
-    Here](https://reilly-lab.github.io/ConversationAlign_LookupDatabaseCreation.html)
+    Here](https://reilly-lab.github.io/ConversationAlign_LookupHowTo_Jun25.html)
     to see how we created this by merging other offline psycholinguistic
     databases into one.  
 
 3.  **Variable Key for ConversationAlign**  
     ConversationAlign currently allows users to compute alignment
     dynamics across \>40 different lexical, affective, and semantic
-    dimensions.[Click
-    Here](https://reilly-lab.github.io/ConversationAlign_VariableLookupKey.pdf)
+    dimensions.\[Click
+    Here\]((<https://reilly-lab.github.io/ConversationAlign_LookupJul25_VariableKey.html>)
     to link to a variable key.  
 
 ## References
