### 2026-01-15 ba89dc8
updated JOSS paper for new guidelines
