+++ paper/paper.md
@@ -4,8 +4,8 @@ tags:
   - R
   - conversation analysis
   - alignment
-  - psycholinguistics
-  - natural language processing
+  - language
+  - discourse
 authors:
   - name: Jamie Reilly
     orcid: 0000-0002-0891-438X
@@ -45,57 +45,30 @@ affiliations:
   - name: Wharton School, University of Pennsylvania, United States
     index: 3
 
-correspondence: jamie.reilly@temple.edu
-date: "27 July 2025"
+date: "15 January 2026"
 bibliography: paper.bib
-abstract: |
- `ConversationAlign` is an R package that executes a series of operations upon one or more conversation transcripts (i.e., two-person dialogues). Transcripts nominally contain at least two variables (speaker identity and text). In addition to these essential fields, `ConversationAlign` will retain all other meta-data such as timestamps, demographics, and grouping variables.  `ConversationAlign` imports raw transcripts into R, appends unique document identifiers, and concatenates all conversations into a single dataframe. `ConversationAlign` generates corpus analytics characterizing the conversation transcript(s) of interest. Users guide a number of text cleaning operations such as stopword removal and lemmatization. The package ultimately vectorizes the original text into a one-word-per-row format. `ConversationAlign` yokes published norms to each content word spanning more than 40 lexical, affective, and semantic dimensions (e.g. word length, morphological complexity, arousal, valence). `ConversationAlign` outputs summary data for each conversation including main effects and indices of local and global alignment for each specified dimension of interest.   
-keywords: |
-  conversation analysis; discourse; language processing; alignment
-authorcontributions: |
-  JR, VU, and BS conceived the software. All authors drafted and edited the paper.
 funding: |
-  This research was funded by R01 DC013063 from the US National Institute on Deafness and Other Communicative Disorders (NIH/NIDCD).
-conflictsofinterest: |
-  The authors declare no conflicts of interest.
+ R01 DC013063 (NIH/NIDCD)
 output: rticles::joss_article 
-keep_tex: true
-csl: apa.csl
-journal: JOSS
 ---
 
+# Summary
+ `ConversationAlign` executes a series of operations upon one or more conversation transcripts (i.e., two-person dialogues).  `ConversationAlign` imports raw language transcripts into R, appends unique document identifiers, and concatenates all conversations into a single dataframe. `ConversationAlign` generates corpus analytics and executes text cleaning operations such as stopword removal and lemmatization. The package vectorizes text and yokes published norms to each content word spanning more than 40 lexical, affective, and semantic dimensions. `ConversationAlign` outputs summary statistics each conversation including main effects and indices of local and global alignment for each specified dimension of interest.   
+ 
+# Software Design
+ `ConversationAlign` was designed as a user-friendly R package with no proprietary components or input from AI (large language models). We were intentional about making the software accessible to users who do not have  extensive backgrounds in computational linguistics or Natural Language Processing.
+
 # Statment of Need
-Although many excellent text analysis applications exist (e.g., Quanteda [@benoit_quanteda_2018] and Korpus [@michalke_korpus_2018]), we know of no R packages that are tailored to the unique demands of conversation analysis (but for Python see ALIGN [@duran_align_2019]). `ConversationAlign` offers a comprehensive text processing pipeline and novel algoirthms for computing linguistic alignment in 2-person dialogues. This software offers standardization and automation advantages that are in great need in a field that has historically relied heavily upon manual coding systems and subjective human judgment.
[…]
