### 2026-01-16 7b9b87b
updated JOSS paper
