### 2026-01-22 #8625 @reilly-lab
Hi @arfon @samhforbes -- Thanks for your help with considering our overhauled submission. I think we've successfully tackled the new JOSS structure. Please let us know if you need anything or we can help in any way.  ~Jamie
