### 2026-01-29 c9f06e6
Updated JOSS with state of the field section
