+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'ConversationAlign: An R package for Computing Linguistic Alignment and Corpus Analytics in Dyadic Conversation Transcripts'
+title: 'ConversationAlign: Computing Linguistic Alignment and Corpus Analytics in Dyadic Conversation Transcripts'
 tags:
   - R
   - conversation analysis
@@ -45,7 +45,7 @@ affiliations:
   - name: Wharton School, University of Pennsylvania, United States
     index: 3
 
-date: "16 January 2026"
+date: "20 January 2026"
 bibliography: paper.bib
 funding: |
  R01 DC013063 (NIH/NIDCD)
@@ -55,15 +55,18 @@ output: rticles::joss_article
 # Summary
  `ConversationAlign` executes a series of operations upon one or more conversation transcripts (i.e., two-person dialogues).  `ConversationAlign` imports raw language transcripts into R, appends unique document identifiers, and concatenates all conversations into a single dataframe. `ConversationAlign` generates corpus analytics and executes text cleaning operations such as stopword removal and lemmatization. The package vectorizes text and yokes published norms to each content word spanning more than 40 lexical, affective, and semantic dimensions. `ConversationAlign` outputs summary statistics each conversation including main effects and indices of local and global alignment for each specified dimension of interest.   
  
-# Software Design
- `ConversationAlign` was designed as a user-friendly R package with no proprietary components or input from AI (large language models). We were intentional about making the software accessible to users who do not have  extensive backgrounds in computational linguistics or Natural Language Processing.
-
 # Statment of Need
-Although many excellent text analysis applications exist (e.g., `Quanteda` [@benoit:2018] and `Korpus` [@michalke:2018]), we know of no R packages that are tailored to the unique demands of conversation analysis (but for Python see ALIGN [@duran:2019]). `ConversationAlign` offers a comprehensive text processing pipeline and novel algoirthms for computing linguistic alignment in 2-person dialogues. This software offers standardization and automation advantages that are in great need in a field that has historically relied heavily upon manual coding systems and subjective human judgment.
+Many excellent text analysis applications exist (e.g., `Quanteda` [@benoit:2018] and `Korpus` [@michalke:2018]). However, few applications are tailored to the unique demands of conversation analysis (but for Python see ALIGN [@duran:2019]). `ConversationAlign` offers a comprehensive text processing pipeline and novel algoirthms for computing linguistic alignment in 2-person dialogues. This software offers standardization and automation advantages that are in great need in a field that has historically relied heavily upon manual coding systems and subjective human judgment.
+
+# State of the Field
+The fundamental challenges involved in conversation analysis involve not only measuring individual behavior but also characterizing the dynamics of alignment between two or more partners. Conversation analysis has historically been undertaken by a variety of fields (e.g., linguistics, psychology). Recent advances in 2-person neuroscience (e.g., hyperscanning) and natural language processing (NLP) (e.g., time series analysis) are driving development of new methods for modeling human interaction and understanding how people synchronize language and brain activity. `ConversationAlign` is poised to make important contributions to measuring and modeling alignment and integrating physiological with linguistic processes as simultaneous time series data.
+ 
+# Software Design
+ `ConversationAlign` was designed as a user-friendly R package with no proprietary components or input from artificial intelligence (e.g., large language models). Our goal was to make the software accessible to users who do not have extensive backgrounds in computational linguistics or Natural Language Processing.
+
 
 # Research Impact Statement
- `ConversationAlign` has supported two peer-reviewed publications to date in the cognitive neuroscience and psychological methods journals `Cortex` [@reilly:2025] and `Behavior Research Methods` (article in press). 
[…]
