### 2026-02-21 #8625 @samhforbes
Hi @reilly-lab I've just been alerted to the new paper in BRM (congrats BTW!).
Can I check with you on the relationship between the two papers? In submission it sounded like BRM was focusing more on the algorithms, though there is description of the software functionality in the version of the paper I can see.
