### 2026-02-21 #8625 @reilly-lab
> Hi [@reilly-lab](https://github.com/reilly-lab) I've just been alerted to the new paper in BRM (congrats BTW!). Can I check with you on the relationship between the two papers? In submission it sounded like BRM was focusing more on the algorithms, though there is description of the software functionality in the version of the paper I can see.

Hi Sam! Thanks for the kind words on the BRM submission. There shouldn't be a lot of overlap between the JOSS submission and BRM. Mostly that paper involved a use case and description(s) of the math behind how the alignment algorithms work along with some broad details about text cleaning.  We intentionally kept the BRM paper less software-oriented and more focused on the language mechanics and rationale(s) for some of the choices we made about stopword removal, etc.. Happy to chat through any questions/concerns.
