### 2026-03-23 #8625 @samhforbes
Hi @cjbarrie, @ellepannitto, @mbod - would you be interested in reviewing this work for JOSS?
