### 2026-04-08 #8625 @samhforbes
Hi @zhanglipku @elenlefoll would either of you be interested in reviewing this package for JOSS?
