### 2026-04-09 #8625 @elenlefoll
> Hi [@zhanglipku](https://github.com/zhanglipku) [@elenlefoll](https://github.com/elenlefoll) would either of you be interested in reviewing this package for JOSS?

Thanks for asking, but I'm afraid I really don't have any capacity right now or for the next few months. Perhaps @jobschepens or @finkelbert might be interested?
