### 2026-04-15 #10416 @samhforbes
Hi all, this is our review thread.

@ellepannitto, @jobschepens please once you've generated your checklist, raise issues in the repo itself, and link back to them here. Feel free to post here if there are any larger points requiring discussion or clarification, and let me know if you need my input on anything.

@reilly-lab while this is going on perhaps you could add in the missing sections above into the paper?
