### 2026-04-16 #10416 @reilly-lab
> Hi all, this is our review thread.
>
> [@ellepannitto](https://github.com/ellepannitto), [@jobschepens](https://github.com/jobschepens) please once you've generated your checklist, raise issues in the repo itself, and link back to them here. Feel free to post here if there are any larger points requiring discussion or clarification, and let me know if you need my input on anything.
>
> [@reilly-lab](https://github.com/reilly-lab) while this is going on perhaps you could add in the missing sections above into the paper?

Many thanks @samhforbes - document updated. All those sections were in there, but I had capitalized the header titles. Now just the first word in the title is capitalized so that bot should see those!
