### 2026-04-25 #10416 @reilly-lab
@samhforbes We just had a dependency hiccup. A package dependency(`YRmisc`) for ConversationAlign was just removed from CRAN.  I am rewriting the scripts that rely on YRmisc` today, but the package might be down momentarily on CRAN.  It should work from with a devtools installation directly from our github repo though probably later today. Will update ASAP. Sorry!
