+++ R/globals.R
@@ -5,6 +5,7 @@ utils::globalVariables(
   c("CA_orig_stops",
     "clean_me",
     "clean_dyads",
+    "cor",
     "df",
     "df_prep",
     "df_read",
@@ -28,6 +29,7 @@ utils::globalVariables(
     "phon_nsyll",
     "reshaped",
     "score",
+    "setNames",
     "SMART_stops",
     "stdev",
     "sumdat",
