+++ vignettes/ConversationAlign_Introduction.Rmd
@@ -85,7 +85,7 @@ Our PsyArXiv preprint describing the method(s) in greater detail is referenced a
 Sacks, B., Ulichney, V., Duncan, A., Helion, C., Weinstein, S., Giovannetti, T., ... Reilly, J. (2025, March 12). *ConversationAlign: Open-Source Software for Analyzing Patterns of Lexical Use and Alignment in Conversation Transcripts*. to read our preprint. It was recently invited for revision at Behavior Rsearch Methods. We will update when/if eventually accepted there! <br>
 
 2. **Variable Key for ConversationAlign** <br>
-``ConversationAlign`` currently allows users to compute alignment dynamics across >40 different lexical, affective, and semantic dimensions.[Click Here]((https://reilly-lab.github.io/ConversationAlign_LookupJul25_VariableKey.html) to link to a variable key. <br>
+``ConversationAlign`` currently allows users to compute alignment dynamics across >40 different lexical, affective, and semantic dimensions.[Click Here to View Lookup Key](https://reilly-lab.github.io/Jamie_LookupVariableKey.html). <br>
 
 # References
 Lewis, David D., et al. (2004) "Rcv1: A new benchmark collection for text categorization research." Journal of machine learning research 5: 361-397.
