### 2026-05-18 33c3e53
Fix tidyselect deprecation warnings: replace contains() with all_of()
- Replace tidyselect::contains(align_var) with all_of(align_var) in dplyr::across()
- Replace tidyselect::contains() in tidyr::pivot_wider() with explicit column names
- Add id_cols parameter to pivot_wider for explicit column specification
- These changes address tidyselect 1.1.0 deprecation warnings about external vectors" -m "Tests still pass: [ FAIL 0 | WARN 10 | SKIP 0 | PASS 5 ]
Note: 6 tidyselect warnings persist from tidyr::pivot_wider() internal code, not our direct usage.
