### 2026-09-02 #10416 @reilly-lab
> 1. `LGPLv3` license updated in the DESCRIPTION file consistent with the license now
