### 2026-09-02 #10416 @reilly-lab
> This is a really great and good-working, well-developed, and validated software package. It can be used to study many different kinds of dynamics in spoken discourse, such as asymmetry across roles, decay over time in alignment patterns, and alignment across various psycholinguistic dimensions.
>
> The package fits very well to JOSS. I have some minor concerns only. I think these are easily addressable and I would therefore recommend acceptance. (I will copy these notes as an issue in the ConversationAlign repository.)
>
> Looking at the licensing of `ConversationAlign`, I think the license situation as it is now could be clarified. Specifically:
>
> 1. The `DESCRIPTION` file declares: `License: LGPL (>= 3)`, while the `LICENSE.md` file and the `README.md` files declare `LGPLv3`.

Thank you for catching this -- updated license GPL-3 (consistent with point #3), updated news, description, license file

> 3. Dependencies. The package imports `textstem` which has not been updated for a long time and is also still licensed under **GPL-2**. Distributing a package under LGPLv3 that imports a GPL-2 package would violate the terms of GPL-2. A suitable alternative would be `udpipe`. Relicensing `ConversationAlign` under `GPL-2` or `GPL-3` is also possible. Another option is to load `textstem` only in functions that require lemmatization.
> 4. Data Licensing. The database `lookup_Jul25` and accompanying datasets are loaded from the online `ConversationAlign_Data` GitHub repository. The data in `lookup_Jul25` aggregates several third-party resources which are not compatible with the LGPLv3. The NRC Emotion Lexicon and NRC VAD Norms (Mohammad, 2018) are licensed for non-commercial purposes and public hosting elsewhere is not permitted. Other datasets were published as academic open-access supplementary materials and also cannot be commercially distributed. I think it may be OK to add a warning in a `DATA_LICENSE.md`, saying that the LGPLv3 license applies only to the software source code and that the psycholinguistic databases have separate restrictions.>
> When looking at the authorships, I noticed some differences between the package and the paper:
>
> 1. Two people are listed as authors but have no commits. Could you add a note on their roles to align with JOSS authorship guidelines?
> 2. In the paper describing the software in _Behavior Research Methods_ (Sacks et al., 2026), there is a larger, 8-person author list. Could you please clarify whether the individuals missing from the JOSS submission contributed to the software or manuscript in a way that warrants their inclusion in the JOSS paper? It seems that one of them is included in the DESCRIPTION as a contributor.
>
> There are also some minor issues that I list here:
>
> 1. There are some typos in the paper. For example: Statment, algoirthms, companionn
> 2. the explanation of the software design in the paper is a bit brief. I would suggest to describe some of the design and representation choices of dialogues, and how AUC and lagged correlations are computed.
> 3. There are no contribution guidelines
> 4. `dplyr`: the `DESCRIPTION` file mentions `dplyr (>= 0.4.3)` but newer dplyr functions are used.
> 5. Automatic downloading files violates CRAN policies. Maybe implement a helper function, e.g., `download_align_data()`, that users can run themselves.
> 6. `read_dyads()` contains some bugs. The column name `talker` is included in the _text_ column check rather than the _speaker_ column check. `Turn` and `MyText` are missing in the renaming block.
>
> AI usage: I used AI (gemini flash) to contextualize public materials, i.e. carry out explorative analyses using the Terry Gross Fresh Air corpus. I am happy to share the resulting Quarto file if desired.
>
> ## Review checklist for [@jobschepens](https://github.com/jobschepens)
> ### AI usage
> * * [x]  I confirm that I have read the [JOSS AI policy](https://joss.theoj.org/about#ai-policy) and understand that:
>
>     * AI may only be used for **non-substantive assistance** (e.g., grammar checking my review text, contextualizing public materials)* **All evaluative decisions**—including scoring, accept/reject recommendations, and assessments of originality, correctness, significance, and ethics compliance—**must be made exclusively by me**, not AI* I bear full responsibility for all judgments, recommendations, and technical claims in my review* I will disclose any AI assistance at the conclusion of my review
>
> ### Conflict of interest
> * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/main/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://github.com/openjournals/joss/blob/main/CODE_OF_CONDUCT.md).
>
> ### General checks
> * [x]  **Repository:** Is the source code for this software available at the https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign?[x]  **License:** Does the repository contain a plain-text LICENSE or COPYING file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?[x]  **Contribution and authorship:** Has the submitting author ([@reilly-lab](https://github.com/reilly-lab)) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?[x]  **Scope and significance:** Does this submission demonstrate clear research impact or credible scholarly significance as described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html)?[x]  **Data sharing:** If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.[x]  **Reproducibility:** If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.[x]  **Human and animal research:** If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.
>
> ### Development history and open-source practice
> * [x]  **Development timeline:** Does the project show evidence of sustained development over time (preferably months or years) rather than rapid, recent code generation?[x]  **Open development:** Was the software developed openly from early stages? For projects with recent public repositories, is there at least six months of public development history with evidence of releases, public issues/pull requests, and ideally external engagement?[x]  **Collaborative effort:** Does the project show evidence of community engagement such as contributions from multiple developers, issues/discussions from external users, or community influence on the development described in the paper?[ ]  **Good practices:** Does the project look like a reusable, non-throwaway research software project that follows good open-source practices (license, documentation, tests and/or verification processes, releases, and clear contribution/support pathways).
>
> ### Functionality
> * [x]  **Installation:** Does installation proceed as outlined in the documentation?[x]  **Functionality:** Have the functional claims of the software been confirmed?[x]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [x]  **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?[x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.[x]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).[x]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?[x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?[ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
>
> ### Software paper
> * [x]  **Summary (required section):** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?[x]  **A statement of need (required section):** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?[x]  **State of the field (required section):** Do the authors describe how this software compares to other commonly-used packages? If related tools exist, do they provide a clear "build vs. contribute" justification explaining their unique scholarly contribution and why alternatives are insufficient?[ ]  **Software Design (required section):** Does the paper include a "Software Design" section that explains the trade-offs weighed, the design/architecture chosen, and why it matters? Is the content compelling and does it demonstrate meaningful design thinking?[x]  **Research Impact Statement (required section):** Does the paper include a "Research Impact Statement" section providing evidence of realized impact (publications, external use, integrations) or credible near-term significance (benchmarks, reproducible materials, community-readiness signals)? Is the evidence presented compelling?[x]  **AI usage disclosure (required section):** Does the paper include an "AI usage disclosure" section that transparently discloses any use of generative AI in the software creation, documentation, or paper authoring?[ ]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?[x]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
