+++ DESCRIPTION
@@ -14,7 +14,7 @@ Authors@R: c(
     person("Gus", "Cooney", role = "ctb"))
 Maintainer: Jamie Reilly <jamie_reilly@temple.edu>
 Description: Imports conversation transcripts into R, concatenates them into a single dataframe appending event identifiers, cleans and formats the text, then yokes user-specified psycholinguistic database values to each word.  'ConversationAlign' then computes alignment indices between two interlocutors across each transcript for >40 possible semantic, lexical, and affective dimensions. In addition to alignment, 'ConversationAlign' also produces a table of analytics (e.g., token count, type-token-ratio) in a summary table describing your particular text corpus.
-License: LGPL (>= 3)
+License: GPL-3
 Encoding: UTF-8
 Depends: 
     R (>= 3.5)
