+++ articles/ConversationAlign_Introduction.md
@@ -56,6 +56,7 @@ Install and load the development version of `ConversationAlign` from
 [GitHub](https://github.com/) using the `devtools` package.
 
 ``` r
+
 # Load SemanticDistance
 library(ConversationAlign)
 ```
@@ -72,33 +73,35 @@ pipeline reshaopes conversation transcripts.
 #### NurseryRhymes
 
 ``` r
+
 knitr::kable(head(NurseryRhymes, 20), format = "simple")
 ```
 
-| Event_ID   | Participant_ID | Text_Raw                                             |
-|:-----------|:---------------|:-----------------------------------------------------|
-| ItsySpider | Yin            | The itsy-bitsy spider climbed up the water spout     |
-| ItsySpider | Maya           | Down came the rain and washed the spider out         |
-| ItsySpider | Yin            | Out came the sun, and dried up all the rain          |
-| ItsySpider | Maya           | And the itsy-bitsy spider climbed up the spout again |
-| ItsySpider | Yin            | The itsy-bitsy spider climbed up the water spout     |
-| ItsySpider | Maya           | Down came the rain and washed the spider out         |
-| ItsySpider | Yin            | Out came the sun, and dried up all the rain          |
-| ItsySpider | Maya           | And the itsy-bitsy spider climbed up the spout again |
-| ItsySpider | Yin            | The itsy-bitsy spider climbed up the water spout     |
-| ItsySpider | Maya           | Down came the rain and washed the spider out         |
-| ItsySpider | Yin            | Out came the sun, and dried up all the rain          |
-| ItsySpider | Maya           | And the itsy-bitsy spider climbed up the spout again |
-| ItsySpider | Yin            | The itsy-bitsy spider climbed up the water spout     |
-| ItsySpider | Maya           | Down came the rain and washed the spider out         |
-| ItsySpider | Yin            | Out came the sun, and dried up all the rain          |
-| ItsySpider | Maya           | And the itsy-bitsy spider climbed up the spout again |
-| ItsySpider | Yin            | The itsy-bitsy spider climbed up the water spout     |
-| ItsySpider | Maya           | Down came the rain and washed the spider out         |
-| ItsySpider | Yin            | Out came the sun, and dried up all the rain          |
-| ItsySpider | Maya           | And the itsy-bitsy spider climbed up the spout again |
+| Event_ID | Participant_ID | Text_Raw |
+|:---|:---|:---|
+| ItsySpider | Yin | The itsy-bitsy spider climbed up the water spout |
+| ItsySpider | Maya | Down came the rain and washed the spider out |
+| ItsySpider | Yin | Out came the sun, and dried up all the rain |
+| ItsySpider | Maya | And the itsy-bitsy spider climbed up the spout again |
+| ItsySpider | Yin | The itsy-bitsy spider climbed up the water spout |
+| ItsySpider | Maya | Down came the rain and washed the spider out |
+| ItsySpider | Yin | Out came the sun, and dried up all the rain |
+| ItsySpider | Maya | And the itsy-bitsy spider climbed up the spout again |
+| ItsySpider | Yin | The itsy-bitsy spider climbed up the water spout |
+| ItsySpider | Maya | Down came the rain and washed the spider out |
+| ItsySpider | Yin | Out came the sun, and dried up all the rain |
+| ItsySpider | Maya | And the itsy-bitsy spider climbed up the spout again |
+| ItsySpider | Yin | The itsy-bitsy spider climbed up the water spout |
+| ItsySpider | Maya | Down came the rain and washed the spider out |
+| ItsySpider | Yin | Out came the sun, and dried up all the rain |
[…]
