+++ llms.txt
@@ -55,6 +55,7 @@ ConversationAlign from [GitHub](https://github.com/) using the
 `devtools` package.
 
 ``` r
+
 # Check if devtools is installed, if not install it
 if (!require("devtools", quietly = TRUE)) {
   install.packages("devtools")
@@ -94,6 +95,7 @@ library(ConversationAlign)
   name  
 
 ``` r
+
 #will search for folder 'my_transcripts' in your current directory
 MyConvos <- read_dyads()
 
@@ -116,23 +118,24 @@ MyConvos2 <- read_dyads(my_path='/MyStuff')
   speaker information.
 
 ``` r
+
 MaryLittleLamb <- read_1file(MaronGross_2013)
 #print first ten rows of header
 knitr::kable(head(MaronGross_2013, 10), format = "pipe")
 ```
 
-| speaker | text                                                                                                                                                                       |
-|:--------|:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| MARON   | I’m a little nervous but I’ve prepared I’ve written things on a piece of paper                                                                                             |
-| MARON   | I don’t know how you prepare I could ask you that - maybe I will But this is how I prepare - I panic                                                                       |
-| MARON   | For a while                                                                                                                                                                |
-| GROSS   | Yeah                                                                                                                                                                       |
-| MARON   | And then I scramble and then I type some things up and then I handwrite things that are hard to read So I can you know challenge myself on that level during the interview |
-| GROSS   | Being self-defeating is always a good part of preparation                                                                                                                  |
-| MARON   | What is?                                                                                                                                                                   |
-| GROSS   | Being self-defeating                                                                                                                                                       |
-| MARON   | Yes                                                                                                                                                                        |
-| GROSS   | Self-sabotage                                                                                                                                                              |
+| speaker | text |
+|:---|:---|
+| MARON | I’m a little nervous but I’ve prepared I’ve written things on a piece of paper |
+| MARON | I don’t know how you prepare I could ask you that - maybe I will But this is how I prepare - I panic |
+| MARON | For a while |
+| GROSS | Yeah |
+| MARON | And then I scramble and then I type some things up and then I handwrite things that are hard to read So I can you know challenge myself on that level during the interview |
+| GROSS | Being self-defeating is always a good part of preparation |
+| MARON | What is? |
+| GROSS | Being self-defeating |
+| MARON | Yes |
+| GROSS | Self-sabotage |
 
   
   
@@ -158,6 +161,7 @@ word. -Retains metadata
[…]

[…]
